Inventory_Check_Service/requirements.txt
flask
pandas
numpy
flask
connexion[flask]
connexion[uvicorn]
sqlalchemy
# connexion
pathlib
connexion[swagger-ui]
sqlalchemy-utils
sqlalchemy-bigquery
logging
time
google-cloud-pubsub==2.16.0
schedule